+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/condition/GeoShapeCondition.java
@@ -23,38 +23,81 @@ import com.spatial4j.core.shape.Shape;
 import com.spatial4j.core.shape.jts.JtsGeometry;
 import com.stratio.cassandra.lucene.IndexException;
 import com.stratio.cassandra.lucene.schema.mapping.GeoPointMapper;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.spatial.SpatialStrategy;
 import org.apache.lucene.spatial.query.SpatialArgs;
-import org.apache.lucene.spatial.query.SpatialOperation;
 
 import java.text.ParseException;
+import java.util.Collections;
 import java.util.List;
 
 /**
- * A {@link Condition} that matches documents containing a shape contained in a certain bounding box.
+ * {@link Condition} that matches documents related to a JTS geographical shape. It is possible to apply a sequence of
+ * {@link GeoTransformation}s to the provided shape to search for points related to the resulting shape.
+ *
+ * The shapes are defined using the <a href="http://en.wikipedia.org/wiki/Well-known_text"> Well Known Text (WKT)</a>
+ * format.
  *
  * @author Andres de la Pena {@literal <adelapena@stratio.com>}
  */
 public class GeoShapeCondition extends SingleMapperCondition<GeoPointMapper> {
 
+    /** The default spatial operation. */
     public static final GeoOperation DEFAULT_OPERATION = GeoOperation.IS_WITHIN;
 
-    /** The list of shape in WKT format. */
-    public final String shape;
+    /** The spatial context to be used. */
+    public static final JtsSpatialContext CONTEXT = JtsSpatialContext.GEO;
+
+    /** The shape. */
+    public final JtsGeometry geometry;
+
+    /** The spatial operation to be applied. */
     public final GeoOperation operation;
+
+    /** The sequence of transformations to be applied to the shape before searching. */
     public final List<GeoTransformation> transformations;
 
+    /**
+     * Constructor receiving the shape, the spatial operation to be done and the transformations to be applied.
+     *
+     * @param boost The boost for this query clause. Documents matching this clause will (in addition to the normal
+     * weightings) have their score multiplied by {@code boost}. If {@code null}, then {@link #DEFAULT_BOOST} is used as
+     * default.
+     * @param field the field name
+     * @param shape the shape in <a href="http://en.wikipedia.org/wiki/Well-known_text"> WKT</a> format
+     * @param operation The spatial operation to be done.  If {@code null}, then {@link #DEFAULT_OPERATION} is used as
+     * default.
+     * @param transformations the sequence of operations to be applied to the specified shape
+     */
     public GeoShapeCondition(Float boost,
                              String field,
                              String shape,
                              GeoOperation operation,
                              List<GeoTransformation> transformations) {
         super(boost, field, GeoPointMapper.class);
-        this.shape = shape;
+        this.geometry = geometryFromWKT(shape);
         this.operation = operation == null ? DEFAULT_OPERATION : operation;
-        this.transformations = transformations;
+        this.transformations = (transformations == null) ? Collections.<GeoTransformation>emptyList() : transformations;
+    }
+
+    /**
+     * Returns the {@link JtsGeometry} represented by the specified WKT text.
+     *
+     * @param string the WKT text
+     * @return the parsed geometry
+     */
+    private static JtsGeometry geometryFromWKT(String string) {
+        if (StringUtils.isBlank(string)) {
+            throw new IndexException("Shape shouldn't be blank");
+        }
+        try {
+            Shape shape = CONTEXT.getWktShapeParser().parse(string);
+            return CONTEXT.makeShape(CONTEXT.getGeometryFrom(shape));
+        } catch (ParseException e) {
+            throw new IndexException(e, "Shape '%s' is not parseable", string);
+        }
     }
 
     /** {@inheritDoc} */
@@ -62,28 +105,16 @@ public class GeoShapeCondition extends SingleMapperCondition<GeoPointMapper> {
     public Query query(GeoPointMapper mapper, Analyzer analyzer) {
 
         SpatialStrategy strategy = mapper.distanceStrategy;
-        JtsSpatialContext context = JtsSpatialContext.GEO;
-
-        // Parse geometry
-        JtsGeometry geometry;
-        try {
-            Shape shape = context.getWktShapeParser().parse(this.shape);
-            geometry = context.makeShape(context.getGeometryFrom(shape));
-        } catch (ParseException e) {
-            throw new IndexException("Shape is not parseable", e);
-        }
 
         // Apply transformations
-        logger.debug("SHAPE {}", geometry);
+        logger.debug("INITIAL SHAPE: {}", geometry);
+        JtsGeometry transformedGeometry = geometry;
         if (transformations != null) {
             for (GeoTransformation transformation : transformations) {
-                geometry = transformation.transform(geometry, context);
-                logger.debug("TRANSFORMED TO {}", geometry);
+                transformedGeometry = transformation.apply(transformedGeometry, CONTEXT);
+                logger.debug("TRANSFORMED TO: {}", transformedGeometry);
             }
         }
-
-        logger.debug("OPERATION {}", operation);
-        logger.debug("SPATIAL OPERATION {}", operation.getSpatialOperation());
 
         // Build query
         SpatialArgs args = new SpatialArgs(operation.getSpatialOperation(), geometry);
@@ -96,6 +127,9 @@ public class GeoShapeCondition extends SingleMapperCondition<GeoPointMapper> {
     /** {@inheritDoc} */
     @Override
     public String toString() {
-        return toStringHelper(this).add("shape", shape).toString();
+        return toStringHelper(this).add("geometry", geometry)
+                                   .add("operation", operation)
+                                   .add("transformations", transformations)
+                                   .toString();
     }
 }