+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/condition/builder/GeoTransformationBuilder.java
@@ -20,7 +20,6 @@ package com.stratio.cassandra.lucene.search.condition.builder;
 
 import com.stratio.cassandra.lucene.search.condition.GeoDistance;
 import com.stratio.cassandra.lucene.search.condition.GeoTransformation;
-import com.stratio.cassandra.lucene.util.Builder;
 import org.apache.commons.lang3.StringUtils;
 import org.codehaus.jackson.annotate.*;
 import org.slf4j.Logger;
@@ -30,27 +29,27 @@ import org.slf4j.LoggerFactory;
  * @author Andres de la Pena {@literal <adelapena@stratio.com>}
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
-@JsonSubTypes({@JsonSubTypes.Type(value = GeoTransformationBuilder.ClipperBuilder.class, name = "clipper"),
-               @JsonSubTypes.Type(value = GeoTransformationBuilder.IdentityBuilder.class, name = "identity")})
+@JsonSubTypes({@JsonSubTypes.Type(value = GeoTransformationBuilder.Buffer.class, name = "buffer"),
+               @JsonSubTypes.Type(value = GeoTransformationBuilder.Copy.class, name = "copy")})
 public interface GeoTransformationBuilder<T extends GeoTransformation> {
 
     T build();
 
     @JsonTypeName("identity")
-    final class IdentityBuilder implements GeoTransformationBuilder<GeoTransformation.Identity> {
+    final class Copy implements GeoTransformationBuilder<GeoTransformation.Copy> {
 
         @JsonCreator
-        public IdentityBuilder() {
+        public Copy() {
         }
 
         @Override
-        public GeoTransformation.Identity build() {
-            return new GeoTransformation.Identity();
+        public GeoTransformation.Copy build() {
+            return new GeoTransformation.Copy();
         }
     }
 
     @JsonTypeName("clipper")
-    final class ClipperBuilder implements GeoTransformationBuilder<GeoTransformation.Clipper> {
+    final class Buffer implements GeoTransformationBuilder<GeoTransformation.Buffer> {
 
         protected static final Logger logger = LoggerFactory.getLogger(GeoTransformationBuilder.class);
 
@@ -63,24 +62,27 @@ public interface GeoTransformationBuilder<T extends GeoTransformation> {
         private String minDistance;
 
         @JsonCreator
-        public ClipperBuilder(@JsonProperty("max_distance") String maxDistance) {
+        public Buffer(@JsonProperty("max_distance") String maxDistance) {
             this.maxDistance = maxDistance;
         }
 
-        public ClipperBuilder setMinDistance(String minDistance) {
+        /**
+         * Sets the min allowed distance.
+         *
+         * @param minDistance the min distance
+         * @return this with the specified min distance
+         */
+        public Buffer setMinDistance(String minDistance) {
             this.minDistance = minDistance;
             return this;
         }
 
+        /** {@inheritDoc}*/
         @Override
-        public GeoTransformation.Clipper build() {
-            logger.debug("MIN STRING {}", minDistance);
-            logger.debug("MAX STRING {}", maxDistance);
-            GeoDistance min = StringUtils.isBlank(minDistance) ? null : GeoDistance.create(minDistance);
-            GeoDistance max = StringUtils.isBlank(maxDistance) ? null : GeoDistance.create(maxDistance);
-            logger.debug("MIN DISTANCE {}", min);
-            logger.debug("MAX DISTANCE {}", max);
-            return new GeoTransformation.Clipper(max, min);
+        public GeoTransformation.Buffer build() {
+            GeoDistance min = StringUtils.isBlank(minDistance) ? null : GeoDistance.parse(minDistance);
+            GeoDistance max = StringUtils.isBlank(maxDistance) ? null : GeoDistance.parse(maxDistance);
+            return new GeoTransformation.Buffer(max, min);
         }
 
     }